feat(module-task): implement House and Vector classes with necessary methods and logic
Created class House with attributes: id, number, address, building type, number of floors, number of apartments, and exploitation term.
Implemented constructor, set(), get(), and show() methods for the House class, along with an additional method for filtering data and displaying it in the console.

Created class Vector of dimension n, declared an array of m objects.
Implemented a method to check if pairs of vectors are collinear or orthogonal, and output the orthogonal and collinear vectors to the console.